// frontend/Nov20-SES1-Java-Angular_Nov20-SES1-Java-Angular-A_GROUP42_FrontEnd-master/src/app/shared/models/order.ts
import { Product } from './product';
import { PaymentThrough } from './payment-through';
import { OrderStatus } from './order-status';

export class Orders {
    customerEmailId: string;
    orderNumber: number;
    orderId: number;
    dateOfOrder: Date;
    product: Product;
    quantity: number;
    totalPrice: number;
    orderStatus: OrderStatus;
    addressId: number;
    paymentThrough: PaymentThrough;
    dataOfDelivery: Date;
    errorMessage: string;
}